requirements.txt
gradio
langchain
langchain-google-genai
google-generativeai
PyPDF2
faiss-cpu
python-dotenv
